#1022 Changing poll options should be forbidden (when running the poll)

+++ components/com_kunena/controllers/topic.php
@@ -284,45 +284,47 @@ class KunenaControllerTopic extends KunenaController {
 			$this->app->enqueueMessage ( $warning, 'notice' );
 		}
 
-		$poll_title = $fields['poll_title'];
-		if ($poll_title !== null) {
-			// Save changes into poll
-			$poll_options = $fields['poll_options'];
-			$poll = $topic->getPoll();
-			if (! empty ( $poll_options ) && ! empty ( $poll_title )) {
-				$poll->title = $poll_title;
-				$poll->polltimetolive = $fields['poll_time_to_live'];
-				$poll->setOptions($poll_options);
-				if (!$topic->poll_id) {
-					// Create a new poll
-					if (!$topic->authorise('poll.create')) {
+		if ( $message->parent == 0 ) {
+			$poll_title = $fields['poll_title'];
+			if ($poll_title !== null) {
+				// Save changes into poll
+				$poll_options = $fields['poll_options'];
+				$poll = $topic->getPoll();
+				if (! empty ( $poll_options ) && ! empty ( $poll_title )) {
+					$poll->title = $poll_title;
+					$poll->polltimetolive = $fields['poll_time_to_live'];
+					$poll->setOptions($poll_options);
+					if (!$topic->poll_id) {
+						// Create a new poll
+						if (!$topic->authorise('poll.create')) {
+							$this->app->enqueueMessage ( $topic->getError(), 'notice' );
+						} elseif (!$poll->save()) {
+							$this->app->enqueueMessage ( $poll->getError(), 'notice' );
+						} else {
+							$topic->poll_id = $poll->id;
+							$topic->save();
+							$this->app->enqueueMessage ( JText::_ ( 'COM_KUNENA_POLL_CREATED' ) );
+						}
+					} else {
+						// Edit existing poll
+						if (!$topic->authorise('poll.edit')) {
+							$this->app->enqueueMessage ( $topic->getError(), 'notice' );
+						} elseif (!$poll->save()) {
+							$this->app->enqueueMessage ( $poll->getError(), 'notice' );
+						} else {
+							$this->app->enqueueMessage ( JText::_ ( 'COM_KUNENA_POLL_EDITED' ) );
+						}
+					}
+				} elseif ($poll->exists()) {
+					// Delete poll
+					if (!$topic->authorise('poll.delete')) {
+						// Error: No permissions to delete poll
 						$this->app->enqueueMessage ( $topic->getError(), 'notice' );
-					} elseif (!$poll->save()) {
+					} elseif (!$poll->delete()) {
 						$this->app->enqueueMessage ( $poll->getError(), 'notice' );
 					} else {
-						$topic->poll_id = $poll->id;
-						$topic->save();
-						$this->app->enqueueMessage ( JText::_ ( 'COM_KUNENA_POLL_CREATED' ) );
+						$this->app->enqueueMessage ( JText::_ ( 'COM_KUNENA_POLL_DELETED' ) );
 					}
-				} else {
-					// Edit existing poll
-					if (!$topic->authorise('poll.edit')) {
-						$this->app->enqueueMessage ( $topic->getError(), 'notice' );
-					} elseif (!$poll->save()) {
-						$this->app->enqueueMessage ( $poll->getError(), 'notice' );
-					} else {
-						$this->app->enqueueMessage ( JText::_ ( 'COM_KUNENA_POLL_EDITED' ) );
-					}
-				}
-			} elseif ($poll->exists()) {
-				// Delete poll
-				if (!$topic->authorise('poll.delete')) {
-					// Error: No permissions to delete poll
-					$this->app->enqueueMessage ( $topic->getError(), 'notice' );
-				} elseif (!$poll->delete()) {
-					$this->app->enqueueMessage ( $poll->getError(), 'notice' );
-				} else {
-					$this->app->enqueueMessage ( JText::_ ( 'COM_KUNENA_POLL_DELETED' ) );
 				}
 			}
 		}
